% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preds_logit_bcfbma_outsamp_with_cate}
\alias{preds_logit_bcfbma_outsamp_with_cate}
\title{Obtain BARTBMA predictions}
\usage{
preds_logit_bcfbma_outsamp_with_cate(
  overall_sum_trees_mu,
  overall_sum_trees_tau,
  overall_sum_mat_mu,
  overall_sum_mat_tau,
  y_arma,
  BIC_weights,
  num_obs,
  a_mu,
  a_tau,
  z,
  test_data,
  test_pihat,
  z_test,
  include_pi2,
  num_propscores,
  num_test_obs,
  lower_prob,
  upper_prob,
  num_cores,
  root_alg_precision,
  maxit,
  eps_f,
  eps_g,
  num_iter,
  include_cate_intervals
)
}
\description{
Obtain BARTBMA predictions
}
